Fix float/double mixing, thanks Chocobo1 for reporting (#7)

+++ butteraugli/butteraugli.cc
@@ -946,18 +946,18 @@ ImageF DiffPrecompute(const ImageF& xyb0, const ImageF& xyb1) {
       } else {
         x2 = x;
       }
-      double minDir =           fabs(row0_in[x] -  row0_in[x2]);
-      minDir = std::min(minDir, fabs(row0_in[x] -  row0_in2[x]));
-      minDir = std::min(minDir, fabs(row0_in[x] -  row0_in2[x2]));
-      minDir = std::min(minDir, fabs(row0_in[x2] - row0_in2[x]));
-      minDir = std::min(minDir, fabs(row1_in[x] -  row1_in[x2]));
-      minDir = std::min(minDir, fabs(row1_in[x] -  row1_in2[x]));
-      minDir = std::min(minDir, fabs(row1_in[x] -  row1_in2[x2]));
-      minDir = std::min(minDir, fabs(row1_in[x2] - row1_in2[x]));
-      double sup0 = (fabs(row0_in[x] - row0_in[x2]) +
-                     fabs(row0_in[x] - row0_in2[x]));
-      double sup1 = (fabs(row1_in[x] - row1_in[x2]) +
-                     fabs(row1_in[x] - row1_in2[x]));
+      float minDir = fabsf(row0_in[x] - row0_in[x2]);
+      minDir = std::min(minDir, fabsf(row0_in[x] - row0_in2[x]));
+      minDir = std::min(minDir, fabsf(row0_in[x] - row0_in2[x2]));
+      minDir = std::min(minDir, fabsf(row0_in[x2] - row0_in2[x]));
+      minDir = std::min(minDir, fabsf(row1_in[x] - row1_in[x2]));
+      minDir = std::min(minDir, fabsf(row1_in[x] - row1_in2[x]));
+      minDir = std::min(minDir, fabsf(row1_in[x] - row1_in2[x2]));
+      minDir = std::min(minDir, fabsf(row1_in[x2] - row1_in2[x]));
+      double sup0 =
+          (fabsf(row0_in[x] - row0_in[x2]) + fabsf(row0_in[x] - row0_in2[x]));
+      double sup1 =
+          (fabsf(row1_in[x] - row1_in[x2]) + fabsf(row1_in[x] - row1_in2[x]));
       static const double mul0 = 0.758790907906;
       static const double mul1 = 1.51115538575;
       row_out[x] = mul0 * std::min(sup0, sup1) + mul1 * minDir;
